Restructured the SBT configuration to play better together with the IDE IntelliJ Idea.
The shared source code directory as been replaced by a SBT project. The two projects for JS and JVM is now depending on the shared project, but as 'Provided'. They are still adding the shared source directory to their 'unmanagedSourceDirectories' list. Including this directory, that is outside the projects base directory, will not take effect in IntelliJ and generate a warning. But IntelliJ will still see the right dependencies to the shared code due to the project dependencies.